Former-commit-id: 4b95b2270b9b511520a4fe5474b4f5308da635ec

+++ dynet/functors.h
@@ -286,12 +286,12 @@ struct FBinaryLogLoss {
     }
     else if (x_true == 0.f) {
       if (x == 1.f) return -1.f * log(DYNET_DEVICE_MIN);
-      else return (x_true - 1.f) * log1p(-x);
+      else return (x_true - 1.f) * log1pf(-x);
     }
     else {
       if (x == 0.f) return -1.f * log(DYNET_DEVICE_MIN);
       else if (x == 1.f) return -1.f * log(DYNET_DEVICE_MIN);
-      else return -1.f * (x_true * log(x) + (1.f - x_true) * log1p(-x));
+      else return -1.f * (x_true * log(x) + (1.f - x_true) * log1pf(-x));
     }
   }
 };
